+++ R/nr-core.R
@@ -122,10 +122,12 @@ nr_duplicate <- function(nr) {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 nr_crop <- function(nr, x, y, w, h) {
   dst <- nr_new(w, h)
-  nr_blit(dst, nr, 
-          0, 0, 
-          x, y, 
-          w, h, respect_alpha = FALSE)
+  nr_blit(dst = dst, src = nr, 
+          x = 0, y = 0, 
+          xsrc = x, ysrc = y, 
+          w = w, h = h, 
+          hjust = 0, vjust = 0,
+          respect_alpha = FALSE)
   dst
 }