+++ src/init.c
@@ -42,6 +42,13 @@ extern SEXP nr_blit_rotozoom_(SEXP dst_, SEXP x_, SEXP y_,
                               SEXP angle_, SEXP sf_,
                               SEXP respect_alpha_);
 
+extern SEXP nr_blit_(SEXP dst_, SEXP x_, SEXP y_, 
+                     SEXP src_, SEXP xsrc_, SEXP ysrc_, 
+                     SEXP w_, SEXP h_, 
+                     SEXP hjust_, SEXP vjust_, 
+                     SEXP angle_, SEXP sf_,
+                     SEXP respect_alpha_);
+
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Converstion
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -86,6 +93,7 @@ static const R_CallMethodDef CEntries[] = {
   {"nr_blit_ortho_"   , (DL_FUNC) &nr_blit_ortho_   , 11},
   {"nr_blit_bulk_"    , (DL_FUNC) &nr_blit_bulk_    ,  3},
   {"nr_blit_rotozoom_", (DL_FUNC) &nr_blit_rotozoom_, 13},
+  {"nr_blit_"         , (DL_FUNC) &nr_blit_         , 13},
   
   {"matrix_to_nr_", (DL_FUNC) &matrix_to_nr_  , 6},